Update README.md
This notebook is a result of a group assignment at Explore Data Science Academy where we have been tasked to develop a climate change sentimet tracker based on historical Twitter data. We then had to showcase our work with a streamlit app and present our findings to a non-technical audience using slides.